Changed compareTo() and method E() to callCompare()
I used @Luke735's already made logic for grabbing the exponent of each node but expanded on it by converting them to integers as me and @Luke735's agreed on this. These would then be compared later in compareTo(). compareTo() was also changed to include the if statements comparing both exponents passed to compareTo(). These if's return an integer depending on the comparison between the exponents. I also added comments explaining the logic written by @Luke735 and comments on the new code in compareTo().

I also changed the E() method to be now named callCompare() and modified the existing code from @Luke735, to create a series of if statements calling compareTo(). Depending on which if is called the specified node is added to a new list called combinedSortedList which will be passed back to the sortPolys() method in main. 

So far the code compiles but will not return the newly combined sorted list. More work is needed on this method and possibly on the compareTo() method.

In summary some code clean up and I put in place the structure and logic of the callCompare() to sort the combined list.